+++ ChangeLog.md
@@ -6,6 +6,7 @@
 * Allow trailing slash in paths
 * Redirect from / to /events/
 * Allow port to be set with the environment variable PORT
+* Allow meetup group to be set with the environment variable MEETUP\_GROUP
 
 ## 1.0.0  -- 2017-11-06